Adding convolution

+++ lenet_sequential/Convolution.h
@@ -5,8 +5,8 @@
 #include <vector>
 
 // Declare the 2D convolution function
-void convolve2D(const std::vector<std::vector<int>>& input, 
-                std::vector<std::vector<int>>& output, 
-                const std::vector<std::vector<int>>& kernel);
+void convolve2D(const std::vector<std::vector<float>>& input, 
+                std::vector<std::vector<float>>& output, 
+                const std::vector<std::vector<float>>& kernel);
 
 #endif // CONVOLUTION_H
